+++ src/infer_compiler_registry/register_diffusers/__init__.py
@@ -1,6 +1,4 @@
-from onediff.infer_compiler.registry import register
-
-
+from onediff.infer_compiler.transform import register
 
 from diffusers.models.attention_processor import Attention, AttnProcessor2_0
 
@@ -14,3 +12,33 @@ torch2oflow_class_map = {
 
 register(package_names=["diffusers"], torch2oflow_class_map=torch2oflow_class_map)
 
+
+from onediff.infer_compiler.transform import transform_mgr
+from onediff.infer_compiler.transform.builtin_transform import proxy_class
+
+
+_ONEFLOW_HAS_REGISTER_RELAXED_TYPE_API = False
+try:
+    from oneflow.framework.args_tree import register_relaxed_type
+
+    _ONEFLOW_HAS_REGISTER_RELAXED_TYPE_API = True
+except:
+    pass
+
+def register_args_tree_relaxed_types():
+    transformers_mocked = False
+    for pkg_name in transform_mgr._torch_to_oflow_packages_list:
+        if "transformers" in pkg_name:
+            transformers_mocked = True
+            break
+
+    if _ONEFLOW_HAS_REGISTER_RELAXED_TYPE_API and transformers_mocked:
+        from transformers.modeling_outputs import BaseModelOutputWithPooling
+        from transformers.models.clip.modeling_clip import CLIPTextModelOutput
+
+        register_relaxed_type(proxy_class(BaseModelOutputWithPooling))
+        register_relaxed_type(proxy_class(CLIPTextModelOutput))
+    else:
+        pass
+
+register_args_tree_relaxed_types()
