--- src/onediff/infer_compiler/torch_to_oflow/mock_comfy/__init__.py
@@ -1,2 +1,0 @@
-from .attention import CrossAttentionPytorch, SpatialTransformer
-from .linear import Linear
