--- src/onediff/infer_compiler/torch_to_oflow/mock_diffusers/__init__.py
@@ -1,1 +1,0 @@
-from .attention_processor_oflow import Attention, AttnProcessor
