--- src/onediff/infer_compiler/torch_to_oflow/mock_diffusers_quant/__init__.py
@@ -1,1 +1,0 @@
-from .quant_diffusion_pipeline import QuantDiffusionPipeline
