+++ src/onediff/infer_compiler/utils/__init__.py
@@ -1,2 +1,1 @@
 from .oneflow_exec_mode import oneflow_exec_mode, oneflow_exec_mode_enabled
-from .args_tree_util import register_args_tree_relaxed_types
