+++ src/onediff/infer_compiler/with_fx_interpreter.py
@@ -1,6 +1,6 @@
 import torch
 from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
-from .torch_to_oflow import map_args, replace_func, ProxySubmodule
+from .transform import map_args, replace_func, ProxySubmodule
 
 class OneFlowInterpreter(torch.fx.Interpreter):
     from torch.fx.node import Argument, Target
